+++ src/routes/RequestsRoute.test.js
@@ -2,6 +2,7 @@ import React from 'react';
 import {
   stringify,
 } from 'query-string';
+import { createIntl, createIntlCache } from 'react-intl';
 
 import {
   render,
@@ -33,6 +34,8 @@ import RequestsRoute, {
   buildHoldRecords,
   getListFormatter,
   getPrintHoldRequestsEnabled,
+  getLastPrintedDetails,
+  getFilteredColumnHeadersMap,
   urls,
   DEFAULT_FORMATTER_VALUE,
 } from './RequestsRoute';
@@ -62,6 +65,7 @@ import {
   REQUEST_OPERATIONS,
   INPUT_REQUEST_SEARCH_SELECTOR,
   ITEM_QUERIES,
+  PRINT_DETAILS_COLUMNS,
 } from '../constants';
 import { historyData } from '../../test/jest/fixtures/historyData';
 
@@ -83,6 +87,15 @@ const testIds = {
 };
 const requestUrl = 'url';
 
+const intlCache = createIntlCache();
+const intl = createIntl(
+  {
+    locale: 'en-US',
+    messages: {},
+  },
+  intlCache
+);
+
 jest.spyOn(React, 'createRef').mockReturnValue(createRefMock);
 jest.spyOn(document, 'getElementById').mockReturnValue(createDocumentRefMock);
 
@@ -101,6 +114,7 @@ jest.mock('../utils', () => ({
   isMultiDataTenant: jest.fn(),
   generateUserName: jest.fn(),
   getRequestUrl: jest.fn(() => requestUrl),
+  extractPickSlipRequestIds: jest.fn(),
 }));
 jest.mock('./utils', () => ({
   ...jest.requireActual('./utils'),
@@ -121,17 +135,25 @@ jest.mock('../components', () => ({
   LoadingButton: jest.fn(() => null),
   PrintButton: jest.fn(({
     onBeforeGetContent,
+    onBeforePrint,
     children,
-  }) => (
-    <div>
-      <button
-        type="button"
-        onClick={onBeforeGetContent}
-      >onBeforeGetContent
-      </button>
-      {children}
-    </div>
-  )),
+  }) => {
+    const handleClick = () => {
+      onBeforeGetContent();
+      onBeforePrint();
+    };
+    return (
+      <div>
+        <button
+          type="button"
+          onClick={handleClick}
+        >
+          PrintButton
+        </button>
+        {children}
+      </div>
+    );
+  }),
   PrintContent: jest.fn(({ printContentTestId }) => <div data-testid={printContentTestId}>PrintContent</div>)
 }));
 jest.mock('../components/RequestsFilters/RequestsFilters', () => ({ onClear }) => {
@@ -150,14 +172,21 @@ jest.mock('../ViewRequest', () => jest.fn());
 jest.mock('../RequestForm', () => jest.fn());
 jest.mock('../components/SinglePrintButtonForPickSlip', () => jest.fn(({
   onBeforeGetContentForSinglePrintButton,
-}) => (
-  <button
-    type="button"
-    data-testid={testIds.singlePrintButton}
-    onClick={onBeforeGetContentForSinglePrintButton}
-  >Print
-  </button>
-)));
+  onBeforePrintForSinglePrintButton,
+}) => {
+  const handleClick = () => {
+    onBeforeGetContentForSinglePrintButton();
+    onBeforePrintForSinglePrintButton(['reqId']);
+  };
+  return (
+    <button
+      type="button"
+      data-testid={testIds.singlePrintButton}
+      onClick={handleClick}
+    >Print
+    </button>
+  );
+}));
 jest.mock('../components/CheckboxColumn/CheckboxColumn', () => jest.fn(({
   toggleRowSelection,
 }) => (
@@ -212,6 +241,11 @@ const userData = {
   requester: {
     personal: {},
   }
+};
+const printDetailsMockData = {
+  count: 11,
+  lastPrintedDate: '2024-08-03T13:33:31.868Z',
+  lastPrintRequester: { firstName: 'firstName', middleName: 'middleName', lastName: 'lastName' },
 };
 
 SearchAndSort.mockImplementation(jest.fn(({
@@ -341,6 +375,7 @@ describe('RequestsRoute', () => {
         postalCode: '12345',
         countryId: 'US',
       },
+      printDetails: printDetailsMockData,
     }
   ];
   const defaultProps = {
@@ -354,6 +389,12 @@ describe('RequestsRoute', () => {
       url: '{{ env.FOLIO_MD_REGISTRY }}/_/proxy/modules',
     },
     mutator: {
+      circulationSettings: {
+        GET: jest.fn(),
+      },
+      savePrintDetails: {
+        POST: jest.fn().mockResolvedValue(),
+      },
       activeRecord: {
         update: jest.fn(),
       },
@@ -423,7 +464,11 @@ describe('RequestsRoute', () => {
         getState: jest.fn(() => ({ okapi: { token: 'token' } })),
       },
       timezone: 'America/New_York',
-      user: {},
+      user: {
+        user: {
+          username: 'rick'
+        }
+      },
     },
     resources: {
       addressTypes: {
@@ -442,6 +487,13 @@ describe('RequestsRoute', () => {
           value: '{"printHoldRequestsEnabled": true}',
         }],
       },
+      circulationSettings: {
+        records: [{
+          value: {
+            enablePrintLog: 'true',
+          }
+        }]
+      },
       currentServicePoint: {},
       patronBlocks: {
         records: [
@@ -453,7 +505,12 @@ describe('RequestsRoute', () => {
       pickSlips: {
         records: [
           {
-            name: 'pick slip',
+            item: {
+              title: 'test title'
+            },
+            request: {
+              requestID: '393030bc-669e-4a41-81e9-3427c25a3b39',
+            }
           }
         ],
       },
@@ -565,11 +622,37 @@ describe('RequestsRoute', () => {
       expect(printContent).toBeInTheDocument();
     });
 
-    it('should trigger "exportCsv"', async () => {
+    it('should trigger "exportCsv" when "Enable view print details" settings is enabled', async () => {
       await userEvent.click(screen.getByRole('button', { name: 'ui-requests.exportSearchResultsToCsv' }));
 
       await waitFor(() => {
-        expect(exportCsv).toBeCalled();
+        expect(exportCsv).toHaveBeenCalled();
+      });
+    });
+
+    it('should trigger "exportCsv" when "Enable view print details" settings is disabled', async () => {
+      const props = {
+        ...defaultProps,
+        resources: {
+          ...defaultProps.resources,
+          circulationSettings: {
+            ...defaultProps.resources.circulationSettings,
+            records: defaultProps.resources.circulationSettings.records.map(record => ({
+              ...record,
+              value: {
+                ...record.value,
+                enablePrintLog: 'false'
+              }
+            }))
+          },
+        }
+      };
+
+      renderComponent(props);
+      await userEvent.click(screen.queryAllByRole('button', { name: 'ui-requests.exportSearchResultsToCsv' })[1]);
+
+      await waitFor(() => {
+        expect(exportCsv).toHaveBeenCalled();
       });
     });
 
@@ -619,15 +702,21 @@ describe('RequestsRoute', () => {
     it('should render print pick slips label', async () => {
       const printPickSlipsLabel = screen.getByText(labelIds.printPickSlips);
 
-      await userEvent.click(screen.getAllByRole('button', { name: 'onBeforeGetContent' })[0]);
+      await userEvent.click(screen.getAllByRole('button', { name: 'PrintButton' })[0]);
 
       expect(printPickSlipsLabel).toBeInTheDocument();
+    });
+
+    it('should trigger "mutator.savePrintDetails.POST"', async () => {
+      await userEvent.click(screen.getAllByRole('button', { name: 'PrintButton' })[0]);
+
+      expect(defaultProps.mutator.savePrintDetails.POST).toHaveBeenCalled();
     });
 
     it('should render print search slips label', async () => {
       const printSearchSlipsLabel = screen.getByText(labelIds.printSearchSlips);
 
-      await userEvent.click(screen.getAllByRole('button', { name: 'onBeforeGetContent' })[1]);
+      await userEvent.click(screen.getAllByRole('button', { name: 'PrintButton' })[1]);
 
       expect(printSearchSlipsLabel).toBeInTheDocument();
     });
@@ -1000,20 +1089,24 @@ describe('RequestsRoute', () => {
     const isPrintableMock = jest.fn(id => id);
     const toggleRowSelectionMock = jest.fn(id => id);
     const onBeforeGetContentForSinglePrintButtonMock = jest.fn(id => id);
+    const onBeforePrintForSinglePrintButtonMock = jest.fn(id => id);
     const listFormatter = getListFormatter(
       {
         getRowURL: getRowURLMock,
         setURL: setURLMock,
       },
       {
+        intl,
         selectedRows: '',
         pickSlipsToCheck: '',
         pickSlipsData: '',
+        isViewPrintDetailsEnabled: true,
         getPrintContentRef: getPrintContentRefMock,
         isPrintableMock,
         pickSlipsPrintTemplate: '',
         toggleRowSelection: toggleRowSelectionMock,
         onBeforeGetContentForSinglePrintButton: onBeforeGetContentForSinglePrintButtonMock,
+        onBeforePrintForSinglePrintButton: onBeforePrintForSinglePrintButtonMock,
       }
     );
     const requestWithData = {
@@ -1040,6 +1133,7 @@ describe('RequestsRoute', () => {
       pickupServicePoint: {
         name: 'name',
       },
+      printDetails: printDetailsMockData,
     };
     const requestWithoutData = {};
 
@@ -1047,6 +1141,7 @@ describe('RequestsRoute', () => {
       it('should render CheckboxColumn', () => {
         const selectedRequest = {};
         const options = {
+          intl,
           selectedRows: [],
           pickSlipsToCheck: [],
           pickSlipsData: {},
@@ -1054,6 +1149,7 @@ describe('RequestsRoute', () => {
           pickSlipsPrintTemplate: jest.fn(),
           toggleRowSelection: jest.fn(),
           onBeforeGetContentForSinglePrintButton: jest.fn(),
+          onBeforePrintForSinglePrintButton: jest.fn(),
         };
         const formatter = getListFormatter({}, options);
         const result = formatter.select(selectedRequest);
@@ -1068,6 +1164,7 @@ describe('RequestsRoute', () => {
       it('should render SinglePrintButtonForPickSlip', () => {
         const selectedRequest = {};
         const options = {
+          intl,
           selectedRows: [],
           pickSlipsToCheck: [],
           pickSlipsData: {},
@@ -1075,6 +1172,7 @@ describe('RequestsRoute', () => {
           pickSlipsPrintTemplate: jest.fn(),
           toggleRowSelection: jest.fn(),
           onBeforeGetContentForSinglePrintButton: jest.fn(),
+          onBeforePrintForSinglePrintButton: jest.fn(),
         };
         const formatter = getListFormatter({}, options);
         const singlePrintButton = formatter.singlePrint(selectedRequest);
@@ -1236,6 +1334,29 @@ describe('RequestsRoute', () => {
         expect(listFormatter.servicePoint(requestWithData)).toBe(requestWithData.pickupServicePoint.name);
       });
     });
+
+    describe('when formatting copies column', () => {
+      it('should return copies for copies column', () => {
+        expect(listFormatter.copies(requestWithData)).toBe(requestWithData.printDetails.count);
+      });
+    });
+
+    describe('when formatting printed column', () => {
+      it('should return last printed details for printed column', () => {
+        getFullName.mockReturnValueOnce('lastName, firstName middleName');
+
+        const expectedFormattedDate = intl.formatDate(requestWithData.printDetails.lastPrintedDate);
+        const expectedFormattedTime = intl.formatTime(requestWithData.printDetails.lastPrintedDate);
+        const expectedOutput =
+        `lastName, firstName middleName ${expectedFormattedDate}${expectedFormattedTime ? ', ' : ''}${expectedFormattedTime}`;
+
+        expect(listFormatter.printed(requestWithData)).toBe(expectedOutput);
+      });
+
+      it('should return empty string', () => {
+        expect(listFormatter.printed(requestWithoutData)).toBe(DEFAULT_FORMATTER_VALUE);
+      });
+    });
   });
 
   describe('getPrintHoldRequestsEnabled', () => {
@@ -1259,6 +1380,49 @@ describe('RequestsRoute', () => {
       expect(getPrintHoldRequestsEnabled({
         records: [],
       })).toBeFalsy();
+    });
+  });
+
+  describe('getLastPrintedDetails', () => {
+    const lastPrintDetails = {
+      lastPrintRequester: { firstName: 'firstName', middleName: 'middleName', lastName: 'lastName' },
+      lastPrintedDate: '2024-08-03T13:33:31.868Z',
+    };
+    beforeEach(() => {
+      getFullName.mockClear().mockReturnValue('lastName, firstName middleName');
+    });
+
+    it('should render getLastPrintedDetails() correctly', () => {
+      expect(getLastPrintedDetails(lastPrintDetails, intl)).toBeTruthy();
+    });
+
+    it('should return the formatted full name and date/time correctly', () => {
+      const printedDetails = getLastPrintedDetails(lastPrintDetails, intl);
+      const expectedFormattedDate = intl.formatDate(lastPrintDetails.lastPrintedDate);
+      const expectedFormattedTime = intl.formatTime(lastPrintDetails.lastPrintedDate);
+      const expectedOutput =
+        `lastName, firstName middleName ${expectedFormattedDate}${expectedFormattedTime ? ', ' : ''}${expectedFormattedTime}`;
+
+      expect(printedDetails).toBe(expectedOutput);
+    });
+  });
+
+  describe('getFilteredColumnHeadersMap', () => {
+    const columnHeaders = [
+      { value: 'patronComments' },
+      { value: PRINT_DETAILS_COLUMNS.COPIES },
+      { value: PRINT_DETAILS_COLUMNS.PRINTED },
+    ];
+
+    it('should render getFilteredColumnHeadersMap() correctly', () => {
+      expect(getFilteredColumnHeadersMap(columnHeaders)).toBeTruthy();
+    });
+
+    it('should return properly filtered column headers', () => {
+      const filteredColumnHeaders = getFilteredColumnHeadersMap(columnHeaders);
+      const expectedColumnHeaders = [{ value: 'patronComments' }];
+
+      expect(filteredColumnHeaders).toEqual(expectedColumnHeaders);
     });
   });