+++ src/utils.js
@@ -242,7 +242,7 @@ export const getNextSelectedRowsState = (selectedRows, row) => {
   return newSelectedRows;
 };
 
-export const convertToSlipData = (source, intl, timeZone, locale, slipName = SLIPS_TYPE.PICK_SLIP) => {
+export const convertToSlipData = (source, intl, timeZone, locale, slipName = SLIPS_TYPE.PICK_SLIP, user = {}) => {
   return source.map(pickSlip => {
     const {
       item = {},
@@ -251,9 +251,12 @@ export const convertToSlipData = (source, intl, timeZone, locale, slipName = SLI
       currentDateTime = null,
     } = pickSlip;
 
+    const { username } = user;
+
     return {
       'staffSlip.Name': slipName,
       'staffSlip.currentDateTime': buildLocaleDateAndTime(currentDateTime, timeZone, locale),
+      'staffSlip.staffUsername': username,
       'requester.firstName': requester.firstName,
       'requester.lastName': requester.lastName,
       'requester.middleName': requester.middleName,
@@ -311,6 +314,7 @@ export const convertToSlipData = (source, intl, timeZone, locale, slipName = SLI
         : request.holdShelfExpirationDate,
       'request.requestID': request.requestID,
       'request.patronComments': request.patronComments,
+      'request.barcodeImage': `<Barcode>${request.requestID}</Barcode>`,
     };
   });
 };